community_lending_mvp/requirements.txt
Django>=4.2
djangorestframework
python-dateutil
celery
redis
twilio
python-dotenv
psycopg2-binary
